Moved bugfix from the Capabilities provider to the Authentication provider
