+++ src/Provider/Capabilities.php
@@ -20,12 +20,7 @@ use SatisPress\Capabilities as Caps;
  * @since 0.3.0
  */
 class Capabilities extends AbstractHookProvider {
-	protected $disable_authentication;
-	
-	public function __construct( bool $disable_authentication = false ) {
-		$this->disable_authentication = $disable_authentication;
-	}
-	
+
 	/**
 	 * Register hooks.
 	 *
@@ -33,7 +28,6 @@ class Capabilities extends AbstractHookProvider {
 	 */
 	public function register_hooks() {
 		add_filter( 'map_meta_cap', [ $this, 'map_meta_cap' ], 10, 4 );
-		add_filter( 'user_has_cap', [ $this, 'user_has_cap' ], 10, 3 );
 	}
 
 	/**
@@ -57,14 +51,5 @@ class Capabilities extends AbstractHookProvider {
 		}
 
 		return $caps;
-	}
-
-	public function user_has_cap( array $allcaps, array $caps, array $args ) {
-		if ( $this->disable_authentication ) {
-			$allcaps[Caps::DOWNLOAD_PACKAGES] = true;
-			$allcaps[Caps::VIEW_PACKAGES] = true;
-		}
-
-		return $allcaps;
 	}
 }
